bqml/bq_regression_dnn/requirements.txt
google-cloud-bigquery==1.25.0